+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [FEATURE] Add `ecs_up` metrics
 * [FEATURE] Add `ecs_cluster_total` metrics
+* [FEATURE] Add `ecs_service_total` metrics
 * [FEATURE] Add `ecs_service_desired_tasks` metrics
 * [FEATURE] Add `ecs_service_pending_tasks` metrics
 * [FEATURE] Add `ecs_service_running_tasks` metrics